// src/app/api/model/page-partnership-opportunity-dto-out.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PageableObject } from './pageable-object';
import { SortObject } from './sort-object';
import { PartnershipOpportunityDtoOut } from './partnership-opportunity-dto-out';


export interface PagePartnershipOpportunityDtoOut { 
    content?: Array<PartnershipOpportunityDtoOut>;
    empty?: boolean;
    first?: boolean;
    last?: boolean;
    number?: number;
    numberOfElements?: number;
    pageable?: PageableObject;
    size?: number;
    sort?: SortObject;
    totalElements?: number;
    totalPages?: number;
}